+++ src/main/java/com/prgrms/mukvengers/global/config/scheduler/SchedulerConfig.java
@@ -1,0 +1,27 @@
+package com.prgrms.mukvengers.global.config.scheduler;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.SchedulingConfigurer;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
+import org.springframework.scheduling.config.ScheduledTaskRegistrar;
+
+@EnableScheduling
+@Configuration
+@ComponentScan(basePackages = "com.prgrms.mukvengers.global.scheduler")
+public class SchedulerConfig implements SchedulingConfigurer {
+
+	private static final int POOL_SIZE = 10;
+
+	@Override
+	public void configureTasks(ScheduledTaskRegistrar taskRegistrar) {
+		ThreadPoolTaskScheduler threadPoolTaskScheduler = new ThreadPoolTaskScheduler();
+		threadPoolTaskScheduler.setPoolSize(POOL_SIZE);
+		threadPoolTaskScheduler.setThreadNamePrefix("scheduled-task-pool-");
+		threadPoolTaskScheduler.initialize();
+
+		taskRegistrar.setTaskScheduler(threadPoolTaskScheduler);
+	}
+}
+
